+++ node/node.h
@@ -473,7 +473,7 @@ private:
 		beam::io::Address m_RemoteAddr; // for logging only
 
 		Block::SystemState::Full m_Tip;
-		uint8_t m_LoginFlags;
+		uint32_t m_LoginFlags;
 
 		uint64_t m_CursorBbs;
 		TxPool::Fluff::Element* m_pCursorTx;
